Step 4-1
Initially I ran into an error "sqlite3.ProgrammingError: SQLite objects created in a thread can only be used in that same thread. The object was created in thread id 7915032576 and this is thread id 6115995648." The solution seemed to be to either create a connection to the database within the same function, or to specify to not check for same thread.

Is there a recommended method, especially with regard to which one is better practice/more secure?

Reference:
https://stackoverflow.com/questions/48218065/objects-created-in-a-thread-can-only-be-used-in-that-same-thread